Refactor view code for proper message in case of wrong data provided

+++ employeesmanagementsystem/employees/views.py
@@ -15,8 +15,10 @@ class EmployeeList(APIView):
     
     def get(self, request, pk):
         employee = Employee.objects.get(pk=pk)
-        serializer = EmployeeSerializer(employee)
-        return Response(serializer.data,status=status.HTTP_200_OK)
+        if employee:
+            serializer = EmployeeSerializer(employee)
+            return Response(serializer.data,status=status.HTTP_200_OK)
+        return Response({"message":"Employee not found!"},status=status.HTTP_404_NOT_FOUND)
     
     def post(self,request):
         serializer = EmployeeSerializer(data=request.data)
@@ -27,13 +29,17 @@ class EmployeeList(APIView):
     
     def put(self, request, pk):
         employee = Employee.objects.get(pk=pk)
-        serializer = EmployeeSerializer(employee, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if employee:
+            serializer = EmployeeSerializer(employee, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        return Response({"message":"Employee not found!"},status=status.HTTP_404_NOT_FOUND)
     
     def delete(self, request, pk):
         employee = Employee.objects.get(pk=pk)
-        employee.delete()
-        return Response({"message":"Deleted!"},status=status.HTTP_204_NO_CONTENT)
+        if employee:
+            employee.delete()
+            return Response({"message":"Deleted!"},status=status.HTTP_204_NO_CONTENT)
+        return Response({"message":"Employee not found!"},status=status.HTTP_404_NOT_FOUND)
